+++ packages/contracts/src/plugins/governance/admin/AdminSetup.sol
@@ -17,7 +17,7 @@ contract AdminSetup is PluginSetup {
     using Clones for address;
 
     /// @notice The address of `Admin` plugin logic contract to be cloned.
-    address private immutable implementation;
+    address private immutable implementation_;
 
     /// @notice Thrown if the admin address is zero.
     /// @param admin The admin address.
@@ -25,7 +25,7 @@ contract AdminSetup is PluginSetup {
 
     /// @notice The constructor setting the `Admin` implementation contract to clone from.
     constructor() {
-        implementation = address(new Admin());
+        implementation_ = address(new Admin());
     }
 
     /// @inheritdoc IPluginSetup
@@ -41,7 +41,7 @@ contract AdminSetup is PluginSetup {
         }
 
         // Clone plugin contract.
-        plugin = implementation.clone();
+        plugin = implementation_.clone();
 
         // Initialize cloned plugin contract.
         Admin(plugin).initialize(IDAO(_dao));
@@ -90,7 +90,7 @@ contract AdminSetup is PluginSetup {
     }
 
     /// @inheritdoc IPluginSetup
-    function getImplementationAddress() external view returns (address) {
-        return implementation;
+    function implementation() external view returns (address) {
+        return implementation_;
     }
 }