+++ packages/contracts/src/plugins/governance/majority-voting/token/TokenVotingSetup.sol
@@ -255,7 +255,7 @@ contract TokenVotingSetup is PluginSetup {
     }
 
     /// @inheritdoc IPluginSetup
-    function getImplementationAddress() external view virtual override returns (address) {
+    function implementation() external view virtual override returns (address) {
         return address(tokenVotingBase);
     }