+++ packages/contracts/test/framework/plugin/plugin-setup-processor.ts
@@ -172,7 +172,7 @@ describe('Plugin Setup Processor', function () {
     const SetupV4 = await smock.mock<PluginUUPSUpgradeableSetupV4Mock__factory>(
       'PluginUUPSUpgradeableSetupV4Mock'
     );
-    setupUV4 = await SetupV4.deploy(await setupUV3.getImplementationAddress());
+    setupUV4 = await SetupV4.deploy(await setupUV3.implementation());
 
     // Deploy PluginCloneableSetupMock
     const SetupC1 = await smock.mock<PluginCloneableSetupV1Mock__factory>(
@@ -303,9 +303,9 @@ describe('Plugin Setup Processor', function () {
 
       const proxy = await pluginFactory
         .attach(plugin)
-        .callStatic.getImplementationAddress();
+        .callStatic.implementation();
 
-      expect(proxy).to.equal(await setup.callStatic.getImplementationAddress());
+      expect(proxy).to.equal(await setup.callStatic.implementation());
     }
   });
 
@@ -1709,11 +1709,7 @@ describe('Plugin Setup Processor', function () {
         )
       )
         .to.be.revertedWithCustomError(psp, 'PluginProxyUpgradeFailed')
-        .withArgs(
-          proxy,
-          await setupUV2.callStatic.getImplementationAddress(),
-          initData
-        );
+        .withArgs(proxy, await setupUV2.callStatic.implementation(), initData);
     });
 
     it('reverts if preparation has not happened yet for update', async () => {
@@ -1940,8 +1936,8 @@ describe('Plugin Setup Processor', function () {
 
       it('points to the V1 implementation', async () => {
         expect(
-          await PluginUV1.attach(proxy).callStatic.getImplementationAddress()
-        ).to.equal(await setupUV1.callStatic.getImplementationAddress());
+          await PluginUV1.attach(proxy).callStatic.implementation()
+        ).to.equal(await setupUV1.callStatic.implementation());
       });
 
       it('initializes the members', async () => {
@@ -2009,8 +2005,8 @@ describe('Plugin Setup Processor', function () {
 
         it('points to the V2 implementation', async () => {
           expect(
-            await PluginUV2.attach(proxy).callStatic.getImplementationAddress()
-          ).to.equal(await setupUV2.callStatic.getImplementationAddress());
+            await PluginUV2.attach(proxy).callStatic.implementation()
+          ).to.equal(await setupUV2.callStatic.implementation());
         });
 
         it('initializes the members', async () => {
@@ -2067,10 +2063,8 @@ describe('Plugin Setup Processor', function () {
 
           it('points to the V3 implementation', async () => {
             expect(
-              await PluginUV3.attach(
-                proxy
-              ).callStatic.getImplementationAddress()
-            ).to.equal(await setupUV3.callStatic.getImplementationAddress());
+              await PluginUV3.attach(proxy).callStatic.implementation()
+            ).to.equal(await setupUV3.callStatic.implementation());
           });
 
           it('initializes the members', async () => {
@@ -2116,8 +2110,8 @@ describe('Plugin Setup Processor', function () {
 
         it('points to the V3 implementation', async () => {
           expect(
-            await PluginUV3.attach(proxy).callStatic.getImplementationAddress()
-          ).to.equal(await setupUV3.callStatic.getImplementationAddress());
+            await PluginUV3.attach(proxy).callStatic.implementation()
+          ).to.equal(await setupUV3.callStatic.implementation());
         });
 
         it('initializes the members', async () => {
@@ -2163,8 +2157,8 @@ describe('Plugin Setup Processor', function () {
 
       it('points to the V2 implementation', async () => {
         expect(
-          await PluginUV2.attach(proxy).callStatic.getImplementationAddress()
-        ).to.equal(await setupUV2.callStatic.getImplementationAddress());
+          await PluginUV2.attach(proxy).callStatic.implementation()
+        ).to.equal(await setupUV2.callStatic.implementation());
       });
 
       it('initializes the members', async () => {
@@ -2221,8 +2215,8 @@ describe('Plugin Setup Processor', function () {
 
         it('points to the V3 implementation', async () => {
           expect(
-            await PluginUV3.attach(proxy).callStatic.getImplementationAddress()
-          ).to.equal(await setupUV3.callStatic.getImplementationAddress());
+            await PluginUV3.attach(proxy).callStatic.implementation()
+          ).to.equal(await setupUV3.callStatic.implementation());
         });
 
         it('initializes the members', async () => {
@@ -2269,8 +2263,8 @@ describe('Plugin Setup Processor', function () {
 
       it('points to the V3 implementation', async () => {
         expect(
-          await PluginUV3.attach(proxy).callStatic.getImplementationAddress()
-        ).to.equal(await setupUV3.callStatic.getImplementationAddress());
+          await PluginUV3.attach(proxy).callStatic.implementation()
+        ).to.equal(await setupUV3.callStatic.implementation());
       });
 
       it('initializes the members', async () => {
@@ -2355,8 +2349,8 @@ async function updateAndValidatePluginUpdate(
   // on the plugin. The below check for this still is not 100% ensuring,
   // As function `upgradeTo` might be called but event `Upgraded`
   // not thrown(OZ changed the logic or name) which will trick the test to pass..
-  const currentImpl = await currentPluginSetup.getImplementationAddress();
-  const newImpl = await newPluginSetup.getImplementationAddress();
+  const currentImpl = await currentPluginSetup.implementation();
+  const newImpl = await newPluginSetup.implementation();
 
   if (currentImpl != newImpl) {
     // ensure that the logic address was also correctly modified on the proxy.
@@ -2364,6 +2358,6 @@ async function updateAndValidatePluginUpdate(
       'PluginUUPSUpgradeable',
       proxy
     );
-    expect(await proxyContract.getImplementationAddress()).to.equal(newImpl);
+    expect(await proxyContract.implementation()).to.equal(newImpl);
   }
 }